Correct task 1

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -13,8 +13,8 @@
  */
 char *_strncat(char *dest, char *src, int n)
 {
-	size_t dest_len = strlen(dest);
-	size_t i;
+	int dest_len = strlen(dest);
+	int i;
 
 	for (i = 0; i < n && src[i] != '\0'; i++)
 		dest[dest_len + i] = src[i];
